requirements.txt
Django==3.1.6
psycopg2==2.8.6
whitenoise
gunicorn
djangorestframework
requests
django-filter
django-cors-headers
pillow